+++ src/store/__tests__/StoreFactory-test.js
@@ -32,6 +32,19 @@ describe('StoreFactory', () => {
     ).not.toBe(storeFactory);
   });
 
+  it('sets the name', () => {
+    const mockName = 'testing';
+    const newDef = storeFactory.defineName(mockName).getDefinition();
+    expect(newDef.name).toBe(mockName);
+  });
+
+  it('compounds the name', () => {
+    const first = 'first';
+    const second = 'second';
+    const newDef = storeFactory.defineName(first).defineName(second).getDefinition();
+    expect(newDef.name).toBe('second(first)');
+  });
+
   it('sets getter', () => {
     const mockGetter = EMPTY_FUNC;
     const newDef = storeFactory.defineGet(mockGetter).getDefinition();